+++ wgpu-types/src/lib.rs
@@ -436,7 +436,7 @@ bitflags::bitflags! {
 /// implementation needs to support more than is needed. You should ideally only request exactly what
 /// you need.
 ///
-/// See also: https://gpuweb.github.io/gpuweb/#dictdef-gpulimits
+/// See also: <https://gpuweb.github.io/gpuweb/#dictdef-gpulimits>
 #[repr(C)]
 #[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
 #[cfg_attr(feature = "trace", derive(Serialize))]
@@ -847,9 +847,9 @@ impl Default for CullMode {
 pub enum PolygonMode {
     /// Polygons are filled
     Fill = 0,
-    /// Polygons are draw as line segments
+    /// Polygons are drawn as line segments
     Line = 1,
-    /// Polygons are draw as points
+    /// Polygons are drawn as points
     Point = 2,
 }
 
@@ -867,13 +867,14 @@ impl Default for PolygonMode {
 pub struct PrimitiveState {
     /// The primitive topology used to interpret vertices.
     pub topology: PrimitiveTopology,
-    /// The format of index buffers for strip topologies. Should be left `None` for non-strip.
+    /// When drawing strip topologies with indices, this is the required format for the index buffer.
+    /// This has no effect on non-indexed or non-strip draws.
     #[cfg_attr(any(feature = "trace", feature = "replay"), serde(default))]
     pub strip_index_format: Option<IndexFormat>,
     /// The face to consider the front for the purpose of culling and stencil operations.
     #[cfg_attr(any(feature = "trace", feature = "replay"), serde(default))]
     pub front_face: FrontFace,
-    /// The fact culling mode.
+    /// The face culling mode.
     #[cfg_attr(any(feature = "trace", feature = "replay"), serde(default))]
     pub cull_mode: CullMode,
     /// Controls the way each polygon is rasterized. Can be either `Fill` (default), `Line` or `Point`
@@ -1368,7 +1369,7 @@ impl TextureFormat {
         let storage = basic | TextureUsage::STORAGE;
         let all_flags = TextureUsage::all();
 
-        // See https://gpuweb.github.io/gpuweb/#texture-format-caps for reference
+        // See <https://gpuweb.github.io/gpuweb/#texture-format-caps> for reference
         let (required_features, sample_type, srgb, block_dimensions, block_size, allowed_usages) =
             match self {
                 // Normal 8 bit textures
@@ -2462,7 +2463,7 @@ pub struct TextureDataLayout {
 
 /// Specific type of a buffer binding.
 ///
-/// WebGPU spec: https://gpuweb.github.io/gpuweb/#enumdef-gpubufferbindingtype
+/// WebGPU spec: <https://gpuweb.github.io/gpuweb/#enumdef-gpubufferbindingtype>
 #[derive(Clone, Copy, Debug, Eq, PartialEq, Hash)]
 #[cfg_attr(feature = "trace", derive(Serialize))]
 #[cfg_attr(feature = "replay", derive(Deserialize))]
@@ -2508,7 +2509,7 @@ impl Default for BufferBindingType {
 
 /// Specific type of a sample in a texture binding.
 ///
-/// WebGPU spec: https://gpuweb.github.io/gpuweb/#enumdef-gputexturesampletype
+/// WebGPU spec: <https://gpuweb.github.io/gpuweb/#enumdef-gputexturesampletype>
 #[derive(Clone, Copy, Debug, Eq, PartialEq, Hash)]
 #[cfg_attr(feature = "trace", derive(Serialize))]
 #[cfg_attr(feature = "replay", derive(Deserialize))]
@@ -2559,7 +2560,7 @@ impl Default for TextureSampleType {
 
 /// Specific type of a sample in a texture binding.
 ///
-/// WebGPU spec: https://gpuweb.github.io/gpuweb/#enumdef-gpustoragetextureaccess
+/// WebGPU spec: <https://gpuweb.github.io/gpuweb/#enumdef-gpustoragetextureaccess>
 #[derive(Clone, Copy, Debug, Eq, PartialEq, Hash)]
 #[cfg_attr(feature = "trace", derive(Serialize))]
 #[cfg_attr(feature = "replay", derive(Deserialize))]
@@ -2579,7 +2580,7 @@ pub enum StorageTextureAccess {
     /// ```
     WriteOnly,
     /// The texture can be both read and written in the shader.
-    /// [`Features::STORAGE_TEXTURE_ACCESS_READ_WRITE`] must be enabled to use this access mode.
+    /// [`Features::TEXTURE_ADAPTER_SPECIFIC_FORMAT_FEATURES`] must be enabled to use this access mode.
     ///
     /// Example GLSL syntax:
     /// ```cpp,ignore
@@ -2591,10 +2592,10 @@ pub enum StorageTextureAccess {
 /// Specific type of a binding.
 ///
 /// WebGPU spec: the enum of
-/// - https://gpuweb.github.io/gpuweb/#dictdef-gpubufferbindinglayout
-/// - https://gpuweb.github.io/gpuweb/#dictdef-gpusamplerbindinglayout
-/// - https://gpuweb.github.io/gpuweb/#dictdef-gputexturebindinglayout
-/// - https://gpuweb.github.io/gpuweb/#dictdef-gpustoragetexturebindinglayout
+/// - <https://gpuweb.github.io/gpuweb/#dictdef-gpubufferbindinglayout>
+/// - <https://gpuweb.github.io/gpuweb/#dictdef-gpusamplerbindinglayout>
+/// - <https://gpuweb.github.io/gpuweb/#dictdef-gputexturebindinglayout>
+/// - <https://gpuweb.github.io/gpuweb/#dictdef-gpustoragetexturebindinglayout>
 #[derive(Clone, Copy, Debug, Eq, PartialEq, Hash)]
 #[cfg_attr(feature = "trace", derive(Serialize))]
 #[cfg_attr(feature = "replay", derive(Deserialize))]
@@ -2628,7 +2629,7 @@ pub enum BindingType {
         /// A filtering sampler can only be used with a filterable texture.
         filtering: bool,
         /// Use as a comparison sampler instead of a normal sampler.
-        /// For more info take a look at the analogous functionality in OpenGL: https://www.khronos.org/opengl/wiki/Sampler_Object#Comparison_mode.
+        /// For more info take a look at the analogous functionality in OpenGL: <https://www.khronos.org/opengl/wiki/Sampler_Object#Comparison_mode>.
         comparison: bool,
     },
     /// A texture binding.
@@ -2655,7 +2656,7 @@ pub enum BindingType {
     /// layout(set=0, binding=0, r32f) uniform image2D myStorageImage;
     /// ```
     /// Note that the texture format must be specified in the shader as well.
-    /// A list of valid formats can be found in the specification here: https://www.khronos.org/registry/OpenGL/specs/gl/GLSLangSpec.4.60.html#layout-qualifiers
+    /// A list of valid formats can be found in the specification here: <https://www.khronos.org/registry/OpenGL/specs/gl/GLSLangSpec.4.60.html#layout-qualifiers>
     StorageTexture {
         /// Allowed access to this texture.
         access: StorageTextureAccess,